+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ------
 
 - Save each aggregate report record as a separate Splunk event
-- Fix IAMP delete action
+- Fix IMAP delete action (issue # 20)
 - Suppress Splunk SSL validation warnings
 - Change default logging level to `WARNING`